+++ src/utils.rs
@@ -2,17 +2,17 @@ use chrono::{Datelike, NaiveDate, Weekday};
 use whatwg_infra::collect_codepoints;
 
 #[inline]
-pub(crate) fn is_valid_month(month: &u8) -> bool {
+pub(crate) fn is_valid_month(month: &u32) -> bool {
 	(1..=12).contains(month)
 }
 
 #[inline]
-pub(crate) fn is_valid_hour(hour: &u8) -> bool {
+pub(crate) fn is_valid_hour(hour: &u32) -> bool {
 	(0..=23).contains(hour)
 }
 
 #[inline]
-pub(crate) fn is_valid_min_or_sec(val: &u8) -> bool {
+pub(crate) fn is_valid_min_or_sec(val: &u32) -> bool {
 	(0..60).contains(val)
 }
 
@@ -21,7 +21,7 @@ pub(crate) fn collect_ascii_digits(s: &str, position: &mut usize) -> String {
 	collect_codepoints(s, position, |c| c.is_ascii_digit())
 }
 
-pub fn max_days_in_month_year(month: u8, year: u32) -> Option<u8> {
+pub fn max_days_in_month_year(month: u32, year: u32) -> Option<u32> {
 	match month {
 		1 | 3 | 5 | 7 | 8 | 10 | 12 => Some(31),
 		4 | 6 | 9 | 11 => Some(30),
@@ -37,22 +37,22 @@ pub fn max_days_in_month_year(month: u8, year: u32) -> Option<u8> {
 }
 
 // https://html.spec.whatwg.org/multipage/common-microsyntaxes.html#weeks
-pub fn week_number_of_year(year: i32) -> Option<u8> {
+pub fn week_number_of_year(year: i32) -> Option<u32> {
 	// We call unwrap() here since `NaiveDate::from_ymd_opt` returns `None` only
 	// if the month/day are out-of-range, which is not possible here since they're hardcoded.
 	let naive_date = NaiveDate::from_ymd_opt(year, 1u32, 1u32).unwrap();
 	let weekday = naive_date.weekday();
 
 	match weekday {
-		Weekday::Thu => Some(53u8),
+		Weekday::Thu => Some(53u32),
 		Weekday::Wed => {
 			if year % 400 == 0 || (year % 4 == 0 && year % 100 != 0) {
-				Some(53u8)
+				Some(53u32)
 			} else {
-				Some(52u8)
+				Some(52u32)
 			}
 		}
-		_ => Some(52u8),
+		_ => Some(52u32),
 	}
 }
